Objdump print refactor (#9)

+++ src/utils.zig
@@ -30,9 +30,9 @@ pub fn intToHex(num: usize) u8 {
     return if (num < 10) ('0' + @as(u8, @intCast(num))) else ('a' + @as(u8, @intCast(num - 10)));
 }
 
-pub fn getNumberDigitCount(comptime T: type, int: T) u32 {
+pub fn getNumberDigitCount(comptime T: type, int: T) u8 {
     const asFloat: f64 = @floatFromInt(int);
-    const numDigits: u32 = @intFromFloat(@floor(@log10(asFloat)) + 1);
+    const numDigits: u8 = @intFromFloat(@floor(@log10(asFloat)) + 1);
     return numDigits;
 }
